Move domain agnostic screen state and screen state changes from domain specific middleware to the reducer
