+++ ModelViewIntentSample/search/src/main/java/com/novoda/movies/mvi/search/domain/SearchReducer.kt
@@ -10,20 +10,37 @@ internal class SearchReducer(
 
     override fun reduce(state: ScreenState, change: ScreenStateChanges): ScreenState =
         when (change) {
-            ScreenStateChanges.ScreenStateInProgress -> ScreenState.Loading(
+            ScreenStateChanges.IndicateProgress -> ScreenState.Loading(
                 queryString = state.queryString
             )
-            is ScreenStateChanges.ScreenStateCompleted -> ScreenState.Content(
+            is ScreenStateChanges.AddResults -> ScreenState.Content(
                 queryString = state.queryString,
                 results = searchResultsConverter.convert(change.results)
             )
-            is ScreenStateChanges.ScreenStateFailed -> ScreenState.Error(
+            is ScreenStateChanges.HandleError -> ScreenState.Error(
                 queryString = state.queryString,
                 throwable = change.throwable
             )
-            is ScreenStateChanges.ScreenStateQueryUpdate -> ScreenState.Content(
+            is ScreenStateChanges.UpdateSearchQuery -> ScreenState.Content(
                 queryString = change.queryString,
                 results = ViewSearchResults()
             )
         }
+}
+
+internal sealed class ScreenState {
+
+    abstract val queryString: String
+
+    data class Content(override val queryString: String, val results: ViewSearchResults) : ScreenState()
+    data class Loading(override val queryString: String) : ScreenState()
+    data class Error(override val queryString: String, val throwable: Throwable) : ScreenState()
+}
+
+sealed class ScreenStateChanges {
+
+    object IndicateProgress : ScreenStateChanges()
+    data class AddResults(val results: SearchResults) : ScreenStateChanges()
+    data class HandleError(val throwable: Throwable) : ScreenStateChanges()
+    data class UpdateSearchQuery(val queryString: String) : ScreenStateChanges()
 }
